macos: add basic GL view and wire up GL context to use it
This creates a new GdkMacosGLView and adds it to the
GdkMacosWindow when a GdkGLContext has been requested.

Currently, a number of features of the GL renderer are not working
and require further tweaking. For example, textures do not seem
to be working correctly (icons, text, and shadows) and more work
is necessary to discover why.